Implement configuration options for RPC
