+++ ProjectLighthouse.Servers.API/Responses/RpcInformation.cs
@@ -1,5 +1,7 @@
 ﻿#nullable disable
 
+using LBPUnion.ProjectLighthouse.Configuration.ConfigurationCategories;
+
 namespace LBPUnion.ProjectLighthouse.Servers.API.Responses;
 
 public class RpcInformation
@@ -8,21 +10,4 @@ public class RpcInformation
     public string PartyIdPrefix { get; set; }
     public UsernameType UsernameType { get; set; }
     public RpcAssets Assets { get; set; }
-}
-
-public class RpcAssets
-{
-    public string PodAsset { get; set; }
-    public string MoonAsset { get; set; }
-    public string RemoteMoonAsset { get; set; }
-    public string DeveloperAsset { get; set; }
-    public string DeveloperAdventureAsset { get; set; }
-    public string DlcAsset { get; set; }
-    public string FallbackAsset { get; set; }
-}
-
-public enum UsernameType
-{
-    Integer = 0,
-    Username = 1,
 }
